+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.4.8
+
+- updated Readmes
+- added menu bar button to toggle classic/embed flow when embedding is active
+
+
 ### 0.4.7
 
 - fixed overlap detection